Moved investigation plan stuff to separate file
